Move stubbing.rb to spec_helper (less indirection)
